casa-agent/requirements.txt
claude-agent-sdk>=0.1.48
honcho-ai>=2.0.0
python-telegram-bot>=20.0
aiohttp>=3.9
apscheduler>=3.10
pyyaml>=6.0
